embrhp: allow specification of EMPinfo via embeddedCode
The embeddedCode content is processed by reshop. This is mainly to do a
syntax check and most importantly save all the UELs, sets and parameters
in a GMD object. The latter is saved as GDX and loaded when reshop is
called as solver. This was, the user can easily use sets and parameters
in the empinfo file.